+++ src/ecs/query/index.js
@@ -1,0 +1,1 @@
+export * from './query.js'
